File: move domain construction code to DomainEditor
